requirements.txt
fastapi
uvicorn
finnhub-python 